[search-cmd-keys-long-gno] Use long for GTID gno in searchCmdKeys chain
Change begGno, endGno, and CmdKeyItem.seq from int to long across the
searchCmdKeys API and storage layer to prevent overflow for large GTID values.

+++ redis/redis-core/src/main/java/com/ctrip/xpipe/redis/core/store/IndexStore.java
@@ -13,7 +13,7 @@ public interface IndexStore {
     void write(ByteBuf byteBuf) throws IOException;
     void rotateFileIfNecessary() throws IOException;
     void openWriter(CommandWriter cmdWriter) throws IOException;
-    List<Pair<Long,  Long>> locateGtidRange(String uuid, int begGno, int endGno) throws IOException;
+    List<Pair<Long,  Long>> locateGtidRange(String uuid, long begGno, long endGno) throws IOException;
     Pair<Long, GtidSet> locateContinueGtidSet(GtidSet request) throws IOException;
     Pair<Long, GtidSet> locateGtidSetWithFallbackToEnd(GtidSet request) throws IOException;
     boolean increaseLost(GtidSet lost, IOSupplier<Boolean> supplier) throws IOException;
